Actualizar ft/uploads

+++ app/routes/edits_routes.py
@@ -1,9 +1,12 @@
 """Rutas para editar y eliminar fotos, álbumes y perfil de usuario."""
 
+import os
+
 from flask import Blueprint, redirect, render_template, url_for, flash, request
 from flask_login import login_required, current_user
 
-from app.models import db, Photo, Album
+from database.db_config import db
+from database.models import Photo, Album
 from app.forms import EditProfileForm
 
 edits = Blueprint("edits", __name__)
@@ -39,31 +42,26 @@ def edit_profile():
     """Editar el perfil del usuario"""
     form = EditProfileForm()
     if form.validate_on_submit():
-        current_user.first_name = form.first_name.data
-        current_user.last_name = form.last_name.data
+        #current_user.first_name = form.first_name.data
+        #current_user.last_name = form.last_name.data
         current_user.email = form.email.data
-        current_user.age = form.age.data
+        #current_user.age = form.age.data
+        current_user.theme_preference = form.theme_preference.data
         if form.password.data:
             current_user.set_password(form.password.data)
         db.session.commit()
         flash("Tu perfil ha sido actualizado.")
         return redirect(url_for("photos.profile"))
 
-    form.first_name.data = current_user.first_name
-    form.last_name.data = current_user.last_name
+    #form.first_name.data = current_user.first_name
+    #form.last_name.data = current_user.last_name
     form.email.data = current_user.email
-    form.age.data = current_user.age
+    #form.age.data = current_user.age
+    form.theme_preference.data = current_user.theme_preference
 
     # Preparar los datos de seguridad
-    security_info = {
-        "last_login_at": current_user.last_login_at,
-        "current_login_at": current_user.current_login_at,
-        "last_login_ip": current_user.last_login_ip,
-        "current_login_ip": current_user.current_login_ip,
-        "login_count": current_user.login_count,
-    }
 
-    return render_template("edit_profile.html", form=form, security_info=security_info)
+    return render_template("edit_profile.html", form=form)
 
 
 @edits.route("/delete_photo/<int:photo_id>", methods=["POST"])
@@ -77,13 +75,22 @@ def delete_photo(photo_id):
     if photo.user_id != current_user.id:
         flash("No tienes permiso para realizar esta acción.")
         return redirect(url_for("photos.timeline_photos"))
-    try:
-        db.session.delete(photo)
-        db.session.commit()
-        flash("Foto eliminada con éxito.")
-    except Exception as e:
-        db.session.rollback()
-        flash("Ocurrió un error al eliminar la foto.")
+    
+    if photo:
+        try:
+            os.remove(photo.path)
+        except FileNotFoundError:
+            print(f"Archivo no encontrado al intentar borrar: {photo.path}")
+            flash("Error en el directorio del usuario, notificar al admin", "error")
+        try:
+            db.session.delete(photo)
+            db.session.commit()
+            flash("Foto eliminada con éxito.")
+        except Exception as e:
+            db.session.rollback()
+            flash("Ocurrió un error al eliminar la foto.")
+    else:
+        flash("Foto no encontrada", "error")
     return redirect(url_for("photos.timeline_photos"))
